Update CONTRIBUTING.md
Thank you for the reply @dougaitken. Will this information do? Our answers are in the [brackets]
# Contributing Guidelines

Please send a Pull Request if you have any additions or changes.

Pull requests will be accepted based on the following:

1. The pull request adheres to the repository's [Code of Conduct](/CODE_OF_CONDUCT.md) [YES WE ARE]
2. If the company is at least partially remote [YES WE ARE]
3. If you add the company in alphabetical order in the list [CloudApp]
4. If you submit the company with a website [https://www.getcloudapp.com]
5. If you submit the company with the regions that are accepted for remote positions (for more details see the instructions in the [example company profile](/company-profiles/example.md#region)) [WORLDWIDE]
6 If you submit a company profile page and link the company name to it (see example [here](/company-profiles/example.md))

Please adhere to the [Contributor Code of Conduct](CODE_OF_CONDUCT.md).

Thanks for contributing!