Update keen-io.md
Added information for Keen (of Keen.io) based on their website, LinkedIn, and Glassdoor. I based my format off the example.md and edited the page in GitHub's text editor. I will try to edit the ReadMe.md in the same request, but am not quite sure what I'm doing with the requests, so I apologize if I mess it up.