Add new company - GEO Jobe (#1114)
* Add GEO Jobe

Added GEO Jobe and filled out its information.

* Update Readme, Add GEO Jobe

Add GEO Jobe and corresponding information to readme.

* Update ReadMe - GEO Jobe

Fixed redirect link for GEO Jobe to point to company profile

* Delete geojobe.md

File named improperly

* Create geo-jobe.md

renamed file to include '-' for the space

* Update ReadMe - rename GEO Jobe profile file

I updated the company profile file for GEO Jobe to have a '-' for the space, so it can pass the check for the pull request.