Create jobs-kea-cloud
kea.
kea builds conversational AI for restaurants. The company was born in the SF Bay, but has been fully-remote since 2021 and we have no intent of going back to the office.
The kea flock is currently 20-50 people internally and we're spread across North America, and Latin America. 
kea is only able to employ folks physically located in North or Latin America currently.
Working hours are flexible and respected. YES - this is a startup (Series A). NO - this is not grind culture.
careers.kea.ai is where you'll find current listings.